Add getConsignmentType query
The front end needs to be able to get the consignment type to determin
which urls can be seen by a user. The only current query which gets it
is for the export and there's a lot of data in that query that we don't
need so I've made a new one.