Split npm and maven deploy.
We're deploying to npm and maven for any changes to this repository.
This means that dependabot updates for package.json deploy a new version
to maven which triggers a load of updates from ScalaSteward for this
library.

This should only deploy to maven if the changes aren't in the ts
directory.

It's not as simple as checking the diff because the deploy job is run
once the changes are already merged to master and the build job which
triggers this deploy is the same because it's triggered when the PR is
closed.

Now, on a PR, a tag is added if only files in the ts directory have
changed and if that's true, then the deploy job doesn't deploy to maven.