Split npm and maven deploy.
We're deploying to npm and maven for any changes to this repository.
This means that dependabot updates for package.json deploy a new version
to maven which triggers a load of updates from ScalaSteward for this
library.

This should only deploy to maven if the changes aren't in the ts
directory.