Add originalFilePath
This is the field that will hold the original version of a redacted
file. The changes to set this field have been made in the API. There
is still an open question about whether we store this somewhere or just
infer it each time but for now, this is working so this can go in.